backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
google-generativeai
google-adk
numpy
python-multipart
python-dotenv
